abstracted text link styles
